requirements.txt
anyio
assemblyai
huggingface-hub
moviepy
numpy 
pillow
ffmepg-python
google-genai